+++ Content.Shared/Chemistry/SharedChemMaster.cs
@@ -56,23 +56,6 @@ namespace Content.Shared.Chemistry
         }
     }
 
-    // Starlight BEGIN
-    [Serializable, NetSerializable]
-    public sealed class ChemMasterReagentCustomAmountButtonMessage : BoundUserInterfaceMessage
-    {
-        public readonly ReagentId ReagentId;
-        public readonly FixedPoint2 Amount;
-        public readonly bool FromBuffer;
-
-        public ChemMasterReagentCustomAmountButtonMessage(ReagentId reagentId, FixedPoint2 amount, bool fromBuffer)
-        {
-            ReagentId = reagentId;
-            Amount = amount;
-            FromBuffer = fromBuffer;
-        }
-    }
-    // Starlight END
-
     [Serializable, NetSerializable]
     public sealed class ChemMasterCreatePillsMessage : BoundUserInterfaceMessage
     {
